moved type class out of IsOfTypeConstraint class
