+++ common/src/main/java/com/scottlogic/deg/common/profile/constraints/atomic/IsOfTypeConstraint.java
@@ -17,12 +17,9 @@
 package com.scottlogic.deg.common.profile.constraints.atomic;
 
 import com.scottlogic.deg.common.profile.Field;
-import com.scottlogic.deg.common.profile.RuleInformation;
+import com.scottlogic.deg.common.profile.Types;
 
-import java.time.OffsetDateTime;
 import java.util.Objects;
-import java.util.Set;
-import java.util.function.Function;
 
 public class IsOfTypeConstraint implements AtomicConstraint {
     public final Field field;
@@ -31,22 +28,6 @@ public class IsOfTypeConstraint implements AtomicConstraint {
     public IsOfTypeConstraint(Field field, Types requiredType) {
         this.field = field;
         this.requiredType = requiredType;
-    }
-
-    public enum Types {
-        NUMERIC(o -> o instanceof Number),
-        STRING(o -> o instanceof String),
-        DATETIME(o -> o instanceof OffsetDateTime);
-
-        private final Function<Object, Boolean> isInstanceOf;
-
-        Types(final Function<Object, Boolean> isInstanceOf) {
-            this.isInstanceOf = isInstanceOf;
-        }
-
-        public boolean isInstanceOf(Object o) {
-            return isInstanceOf.apply(o);
-        }
     }
 
     @Override
