+++ generator/src/main/java/com/scottlogic/deg/generator/fieldspecs/NumericRestrictionsMergeOperation.java
@@ -19,7 +19,7 @@ package com.scottlogic.deg.generator.fieldspecs;
 import com.google.inject.Inject;
 import com.scottlogic.deg.generator.restrictions.*;
 
-import static com.scottlogic.deg.common.profile.constraints.atomic.IsOfTypeConstraint.Types.NUMERIC;
+import static com.scottlogic.deg.common.profile.Types.NUMERIC;
 
 public class NumericRestrictionsMergeOperation implements RestrictionMergeOperation {
     private final NumericRestrictionsMerger merger;
