+++ generator/src/main/java/com/scottlogic/deg/generator/fieldspecs/StringRestrictionsMergeOperation.java
@@ -18,7 +18,7 @@ package com.scottlogic.deg.generator.fieldspecs;
 
 import com.scottlogic.deg.generator.restrictions.*;
 
-import static com.scottlogic.deg.common.profile.constraints.atomic.IsOfTypeConstraint.Types.STRING;
+import static com.scottlogic.deg.common.profile.Types.STRING;
 
 public class StringRestrictionsMergeOperation implements RestrictionMergeOperation {
     private static final StringRestrictionsMerger stringRestrictionsMerger = new StringRestrictionsMerger();
